Update model booking validations

+++ db/migrate/20230419090150_create_bookings.rb
@@ -3,7 +3,7 @@ class CreateBookings < ActiveRecord::Migration[7.0]
     create_table :bookings do |t|
       t.datetime :start_at
       t.datetime :end_at
-      t.integer :total_price
+      t.float :total_price
       t.references :user, null: false, foreign_key: true
       t.references :service, null: false, foreign_key: true
